feat: Introduce mode management and disable Isis mode
Adds a new 'modes' section to settings and implements logic to disable the Isis mode. This includes updating UI elements and user redirection for managers when Isis mode is turned off. Managers will be automatically switched to 'seishat' mode if they are in 'isis' mode and Isis mode is disabled. Admins and managers can now control Isis mode visibility and functionality.