feat(settings): Allow editing associates without password
Introduces conditional validation for the password field in the AssociateModal.
When editing an existing associate, the password field is no longer mandatory by default. It will only be required if the user actively chooses to change the password. This prevents accidental data loss or unintended password resets during associate updates.